fix: moving resident children that are also entering new relationships
when a parent is entering a new cohabit and marriage relationship he/she will also ask the resident children to join. However, if any of the resident children that is also entering a partner relationship in at the same time as the parent will also be asked to move. This causes the child and its new partner to be in two different households.